Fix Infinite worker robot cargo size tech icon and level (#52)
* Fix Infinite worker robot cargo size tech icon

The icon is missing the stack icon in the bottom right corner. This is how the icon is defined in the vanilla prototype. Tested it and it worked flawlessly.

* Fix Infinite worker robot cargo size tech level

Right now "worker robot cargo size 3" is followed by "Infinite worker robot cargo size 1" which is quite confusing. This change will make it continue at level 4.

* Migration of worker robots tech

* Make use of the vanilla localisation

* Own localisation no longer needed

* Adding missing file extension #52

Migration doesn't run without file extension

* Updating changelog #52

Co-authored-by: KiwiHawk <59639+KiwiHawk@users.noreply.github.com>