Add fix for Space Age crafting changes
Because Space Age loads after Bob's Classes, the various crafting categories that are added to characters are not carried over to the new character classes that are created. There are several ways that this could be fixed, but updating crafting_categories in the data-updates stage like I have here should be the most robust overall solution.