Bob's Modules 2.0 (#199)
* Update module-merged.lua to 2.0

Removed the "bonus" variable in the module effects, since it caused crashes in Factorio 2.0

* Update module.lua to Factorio 2.0

Removed the "bonus" variable in the module effects, since it caused a crash in Factorio 2.0

* Update data-updates.lua (Changed research internal name)

Replaced advanced-electronics-2 with processing-unit since the name of the research changed in the base mod

* Update beacon.lua research requirement

Replaced advanced-electronics-2 with processing-unit since the name of the research changed in the base mod

* Update beacon-updates.lua research requirements

Replaced advanced-electronics-2 with processing-unit since the name of the research changed in the base mod

* Update electronics-updates.lua research requirements

Replaced advanced-electronics-2 with processing-unit since the name of the research changed in the base game

* Update module.lua research requirements

Replaced advanced-electronics-2 with processing-unit since the name of the research changed in the base game

* Update module-merged.lua adding quality

Added quality as a bonus for god modules. Might be used to create Bob's quality modules

* Update module-merged.lua

Changed formatting and corrected typos

* Update data.lua changed productivity option

Update 2.0 added a productivity limitation to the game itself.  This option made the default limitation of Bob's Modules non-functional. This commit changes the mod to simply remove the limitation if the option is turned off.

* Update settings.lua add quality

Added quality bonus in order to be used for the god modules, and prepares the terrain for actual quality bonuses

* Update bobmodules.cfg

Added Quality modules to the english version

* Update bobmodules.cfg

Added quality to the russian version.

Courtesy of google translate.

* Update settings.lua added quality option

Added an option to allow beacons to transmit quality.

* Update data-updates.lua added quality to beacons

Allowed beacons to accept quality modules (Including god module)

* Update bobmodules.cfg

Added the english version for the option menu

* Update bobmodules.cfg adding quality

Updated the russian version to include the transmission of quality.

I used google translate, so it might be inaccurate.

* Update data-updates.lua

Removed the ability to add quality modules on beacons that I added on the productivity option. It now has its own, separate, option.

* Update settings.lua disabled quality option without the dlc

Disabled the quality transmission if the "quality" module isn't enabled.

* Update data-updates.lua disabled quality spreading when the quality module isn't enabled

disabled quality spreading when the quality module isn't enabled

* Update data-updates.lua

Removed the line I forgot to change 2 months ago that enabled quality spreading on productivity spreading. I was sure I did it, turns out I didn't.

Fixed now.

* Only add quality settings if Quality mod is enabled

* Fixing startup error

* Removing unused property

* Fixing missing recipe locale

* Fix god module quality

* Don't add quality to God Modules else they can't be used in recipes which don't accept quality

Maybe we could create a separate item Quality God Module?

* Bob's Modules rebalance #269

* Change god modules to have -80% energy consumption

* Delete unused icons

* Add quality module icons

* Make all beacons share effect scaling

* Change efficiency scaling

-40 to -80

* Change gem usage so all gems are used and each module type is different

* Module recipe rebalancing

* Fixing missing prerequisite

---------

Co-authored-by: KiwiHawk <59639+KiwiHawk@users.noreply.github.com>