Classes: fix getting kicked out of cargo pod
Resolves #387 

Since this is something that happens when you switch characters while in any vehicle, I doubt there's any way to prevent it from happening directly. But a clean and simple solution is to simply block on_gui_click from doing anything in this mod while you are in a Cargo Pod.