Sync "bob-resource" subgroup (#496)
This brings the "bob-resource" subgroup in bobelectronics in line with the other two places where it is defined. Those being bobplates and bobwarfare.

In the case where you have bobplates and bobelectronics active, but not bobwarfare, this subgroup would be moved to an odd place within the Intermediates tab instead of the Materials tab because the bobelectronics definition overrides the bobplates definition due to coming later.