Classes: fix getting kicked out of cargo pod (#480)
* Classes: fix getting kicked out of cargo pod

Resolves #387 

Since this is something that happens when you switch characters while in any vehicle, I doubt there's any way to prevent it from happening directly. But a clean and simple solution is to simply block on_gui_click from doing anything in this mod while you are in a Cargo Pod.

* Updated changelog

---------

Co-authored-by: KiwiHawk <59639+KiwiHawk@users.noreply.github.com>